+++ src/sqlalchemy_cratedb/__init__.py
@@ -26,6 +26,7 @@ from .sa_version import SA_1_4, SA_VERSION
 from .support import insert_bulk
 from .type.array import ObjectArray
 from .type.geo import Geopoint, Geoshape
+from .type.ip import IP
 from .type.object import ObjectType
 from .type.vector import FloatVector, knn_match
 
@@ -72,6 +73,7 @@ __all__ = [
     FloatVector,
     Geopoint,
     Geoshape,
+    IP,
     ObjectArray,
     ObjectType,
     UnresolvedType,