+++ src/sqlalchemy_cratedb/type/__init__.py
@@ -1,11 +1,13 @@
 from .array import ObjectArray
 from .geo import Geopoint, Geoshape
+from .ip import IP
 from .object import ObjectType
 from .vector import FloatVector, knn_match
 
 __all__ = [
     Geopoint,
     Geoshape,
+    IP,
     ObjectArray,
     ObjectType,
     FloatVector,